♻️ refactor: 쿼리 dsl 메서드 위치 변경
